refactor(layout): types/+errors/ fold into utils/, utils/ai promoted to service/ai
