+++ src/commands/doctor/__tests__/doctor.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import { doctor, formatDoctor } from '../index.js';
 import type { ProfileRepository } from '../../../repository/profileRepository.js';
-import type { Profile } from '../../../types/index.js';
+import type { Profile } from '../../../utils/types/index.js';
 import type { AppContext } from '../../../cli/appContext.js';
 
 // Minimal AppContext stub — doctor() only ever touches profileRepository.
