+++ src/commands/status/index.ts
@@ -1,5 +1,5 @@
 import { createAppContext, type AppContext } from '../../cli/appContext.js';
-import type { StatusOptions } from '../../types/commands.js';
+import type { StatusOptions } from '../../utils/types/commands.js';
 import type { StatusSummary } from '../../application/statusApplicationService.js';
 
 /**
