+++ src/commands/tailor/index.ts
@@ -1,4 +1,4 @@
-import type { TailorOptions, TailorResult } from '../../types/index.js';
+import type { TailorOptions, TailorResult } from '../../utils/types/index.js';
 import type { AnalyzeResult, WriteStepResult } from '../../application/tailorApplicationService.js';
 import { createAppContext, type AppContext } from '../../cli/appContext.js';
