+++ src/service/ai/registry.ts
