+++ src/service/ai/types.ts
