+++ src/utils/types/job.ts
